# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KeyStone AD Management API",
        "description": "API for managing on-premise Active Directory (Multi-Domain)",
        "contact": {
            "name": "Keystone Admin",
            "email": "admin@keystone.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "List users",
                "operationId": "3ab81e3f9b24b4f36ab08d72349ac632",
                "parameters": [
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "Domain to search (e.g., ncc.lab)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nameFilter",
                        "in": "query",
                        "description": "Filter by display name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statusFilter",
                        "in": "query",
                        "description": "Filter by account status (true=enabled, false=disabled)",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "hasAdminAccount",
                        "in": "query",
                        "description": "Filter by admin account existence",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Create a new Active Directory user",
                "operationId": "ba548b97148bb5c0d8cab3bc3f128bfd",
                "requestBody": {
                    "description": "User creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{samAccountName}": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Get user details",
                "operationId": "451b60d047cde9dbffe1be188a1b2ab0",
                "parameters": [
                    {
                        "name": "samAccountName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "Domain to search (e.g., ncc.lab)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "Update a user's details",
                "operationId": "2534af74d19ad5c12f564c7834c03883",
                "parameters": [
                    {
                        "name": "samAccountName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{samAccountName}/enable": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Enable a user account",
                "operationId": "b26a8817de6c8140a93be568ff5eb433",
                "parameters": [
                    {
                        "name": "samAccountName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "domain": {
                                        "type": "string",
                                        "example": "ncc.lab"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Account enabled"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{samAccountName}/disable": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Disable a user account",
                "operationId": "2d933859e8d3f260dc193d3dcf292a61",
                "parameters": [
                    {
                        "name": "samAccountName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "domain": {
                                        "type": "string",
                                        "example": "ncc.lab"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Account disabled"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{samAccountName}/unlock": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Unlock a user account",
                "operationId": "6378b893bb20244041d3f34089cddb91",
                "parameters": [
                    {
                        "name": "samAccountName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "domain": {
                                        "type": "string",
                                        "example": "ncc.lab"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Account unlocked"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateUserRequest": {
                "title": "Create User Request",
                "required": [
                    "domain",
                    "samAccountName",
                    "firstName",
                    "lastName"
                ],
                "properties": {
                    "domain": {
                        "description": "Domain to create the user in (e.g., ncc.lab)",
                        "type": "string"
                    },
                    "samAccountName": {
                        "description": "User's sAMAccountName (login name)",
                        "type": "string"
                    },
                    "firstName": {
                        "description": "User's first name",
                        "type": "string"
                    },
                    "lastName": {
                        "description": "User's last name",
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "description": "User's date of birth (YYYY-MM-DD)",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "mobileNumber": {
                        "description": "User's mobile number",
                        "type": "string",
                        "nullable": true
                    },
                    "createAdminAccount": {
                        "description": "Create a privileged '-a' account for this user",
                        "type": "boolean",
                        "default": false
                    },
                    "optionalGroupsForStandardUser": {
                        "description": "List of *standard* groups to add the user to.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "optionalGroupsForHighPrivilegeUsers": {
                        "description": "List of *privileged* groups to add the *admin* account to (if createAdminAccount is true).",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateUserRequest": {
                "title": "Update User Request",
                "required": [
                    "domain"
                ],
                "properties": {
                    "domain": {
                        "description": "Domain of the user (e.g., ncc.lab)",
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "description": "User's date of birth (YYYY-MM-DD)",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "mobileNumber": {
                        "description": "User's mobile number",
                        "type": "string",
                        "nullable": true
                    },
                    "hasAdminAccount": {
                        "description": "Manage the privileged '-a' account for this user (requires L3)",
                        "type": "boolean"
                    },
                    "optionalGroups": {
                        "description": "List of *standard* groups the user should be a member of. Groups not in this list will be removed.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}